Add files via upload
This servo3way file has a few issues in reference to the design I'm attempting to accomplish.  It turns the motors in order of the code, I.E the progression of the motors movement are dependent on the motor before them.  A good starting point, but need to figure out how to make the servo move independently of one another and stop moving at a specified range of motion.